feat: add direct font buffer conversion option
