+++ .changeset/direct-font-buffers.md
@@ -1,0 +1,5 @@
+---
+"pptx-glimpse": minor
+---
+
+Add a browser-safe `fonts` conversion option that accepts `ArrayBuffer` or `Uint8Array` font data directly for SVG and PNG rendering.
