+++ packages/renderer/src/node.ts
@@ -1,0 +1,2 @@
+export { createOpentypeSetupFromSystem } from "./font/opentype-system-helpers.js";
+export { collectFontFilePaths } from "./font/system-font-loader.js";
