+++ packages/renderer/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/node.ts"],
   format: ["cjs", "esm"],
   dts: true,
   clean: true,
