+++ tsconfig.json
@@ -18,7 +18,8 @@
     "paths": {
       "@pptx-glimpse/document": ["packages/document/src/index.ts"],
       "@pptx-glimpse/editor-core": ["packages/editor-core/src/index.ts"],
-      "@pptx-glimpse/renderer": ["packages/renderer/src/index.ts"]
+      "@pptx-glimpse/renderer": ["packages/renderer/src/index.ts"],
+      "@pptx-glimpse/renderer/node": ["packages/renderer/src/node.ts"]
     }
   },
   "include": ["packages/*/src/**/*"],
